Add CMI integration and update related localization files
Added integration with the CMI plugin to allow handling of commands and enhance usability. Updated localization files for multiple languages to reflect changes in the plugins integrated with the system. Also, refactored AchievePlayTimeRunnable to support and take CMI's AFK status into account. The AFK status of a player is now considered when updating their playtime, preventing an increase in playtime when a player is marked as AFK in either EssentialsX or CMI.

Added CMILib as a dependency so that we can slowly convert AdvancedAchievements to use classes and methods from it.